refector: 소켓 로직과 게임 룸 생성, 수정, 삭제와 관련된 로직을 서로 분리하였음
